Revert "Merge branch 'master' into A-Packages"
This reverts commit 96cfb1505db2cf43be15eb6cab8ba86e9d33ab62, reversing
changes made to 89e2f36d076040b1b937e90e166224aeec582323.
