+++ src/main/java/echolex/command/Command.java
@@ -1,3 +1,8 @@
+package echolex.command;
+
+import echolex.error.EchoLexException;
+import echolex.task.*;
+
 import java.time.LocalDateTime;
 import java.util.HashMap;
