+++ src/main/java/echolex/command/Parser.java
@@ -1,3 +1,7 @@
+package echolex.command;
+
+import echolex.error.EchoLexException;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
