+++ src/main/java/echolex/error/EchoLexException.java
@@ -1,3 +1,5 @@
+package echolex.error;
+
 public class EchoLexException extends Exception {
     public EchoLexException() {
         super("EchoLex exception occurred");
