+++ src/main/java/echolex/task/Event.java
@@ -1,3 +1,5 @@
+package echolex.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
