+++ src/main/java/echolex/task/Task.java
@@ -1,3 +1,5 @@
+package echolex.task;
+
 enum Importance {
     LOW,
     MEDIUM,
