+++ src/main/java/echolex/task/TaskList.java
@@ -1,3 +1,5 @@
+package echolex.task;
+
 import java.util.ArrayList;
 
 public class TaskList extends ArrayList<Task> {
