+++ src/main/java/echolex/task/Todo.java
@@ -1,3 +1,5 @@
+package echolex.task;
+
 public class Todo extends Task {
     
     public Todo(String description, Boolean isDone) {
